+++ library/QATools/QATools/PageObject/ElementLocator/WaitingElementLocator.php
@@ -15,6 +15,7 @@ use Behat\Mink\Element\NodeElement;
 use QATools\QATools\PageObject\Annotation\TimeoutAnnotation;
 use QATools\QATools\PageObject\ISearchContext;
 use QATools\QATools\PageObject\Property;
+use QATools\QATools\PageObject\SeleniumSelector;
 
 /**
  * Class, that locates WebElements that might not be present at the moment.
@@ -34,12 +35,16 @@ class WaitingElementLocator extends DefaultElementLocator
 	/**
 	 * Creates a new element locator.
 	 *
-	 * @param Property       $property       Property.
-	 * @param ISearchContext $search_context The context to use when finding the element.
+	 * @param Property         $property          Property.
+	 * @param ISearchContext   $search_context    The context to use when finding the element.
+	 * @param SeleniumSelector $selenium_selector Selenium selector.
 	 */
-	public function __construct(Property $property, ISearchContext $search_context)
-	{
-		parent::__construct($property, $search_context);
+	public function __construct(
+		Property $property,
+		ISearchContext $search_context,
+		SeleniumSelector $selenium_selector
+	) {
+		parent::__construct($property, $search_context, $selenium_selector);
 
 		/** @var TimeoutAnnotation[] $annotations */
 		$annotations = $property->getAnnotationsFromPropertyOrClass('@timeout');