+++ tests/QATools/QATools/PageObject/ElementLocator/DefaultElementLocatorFactoryTest.php
@@ -15,6 +15,7 @@ use Mockery as m;
 use Mockery\Adapter\Phpunit\MockeryPHPUnitIntegration;
 use PHPUnit\Framework\TestCase;
 use QATools\QATools\PageObject\ElementLocator\DefaultElementLocatorFactory;
+use QATools\QATools\PageObject\SeleniumSelector;
 
 class DefaultElementLocatorFactoryTest extends TestCase
 {
@@ -32,8 +33,9 @@ class DefaultElementLocatorFactoryTest extends TestCase
 
 	public function testCreateLocator()
 	{
+		$selenium_selector = m::mock(SeleniumSelector::class);
 		$search_context = m::mock('\\QATools\\QATools\\PageObject\\ISearchContext');
-		$factory = new DefaultElementLocatorFactory($search_context);
+		$factory = new DefaultElementLocatorFactory($search_context, $selenium_selector);
 
 		$property = m::mock(self::PROPERTY_CLASS);
 		$property->shouldReceive('getAnnotationsFromPropertyOrClass')->with('@timeout')->once()->andReturn(array());